Adding to the product _MUST BE ATOMIC_
Since we are adding to the product in _global memory_ it's very
important that the add operating happens atomically. If we disregard
this and attempt to do a normal `+=` operation then the behavior will be
undefined due to a race condition on accessing the memory and then
setting it.

Luckily CUDA offers a very nice attomicAdd function that we can call
passing a memory location and the value to add to that memory.